ggml-*: move the ggml interfaces to a dedicated file
